// frontend/web/src/services/API/proserve-wb-packaging-api/models/GetRecipeVersionsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { RecipeVersion } from './RecipeVersion';
import {
    RecipeVersionFromJSON,
    RecipeVersionFromJSONTyped,
    RecipeVersionToJSON,
} from './RecipeVersion';

/**
 * 
 * @export
 * @interface GetRecipeVersionsResponse
 */
export interface GetRecipeVersionsResponse {
    /**
     * 
     * @type {Array<RecipeVersion>}
     * @memberof GetRecipeVersionsResponse
     */
    recipeVersions: Array<RecipeVersion>;
}

/**
 * Check if a given object implements the GetRecipeVersionsResponse interface.
 */
export function instanceOfGetRecipeVersionsResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "recipeVersions" in value;

    return isInstance;
}

export function GetRecipeVersionsResponseFromJSON(json: any): GetRecipeVersionsResponse {
    return GetRecipeVersionsResponseFromJSONTyped(json, false);
}

export function GetRecipeVersionsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetRecipeVersionsResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'recipeVersions': ((json['recipe_versions'] as Array<any>).map(RecipeVersionFromJSON)),
    };
}

export function GetRecipeVersionsResponseToJSON(value?: GetRecipeVersionsResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'recipe_versions': ((value.recipeVersions as Array<any>).map(RecipeVersionToJSON)),
    };
}