// frontend/web/src/services/API/proserve-wb-packaging-api/models/RecipeComponentVersion.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RecipeComponentVersion
 */
export interface RecipeComponentVersion {
    /**
     * Component ID
     * @type {string}
     * @memberof RecipeComponentVersion
     */
    componentId: string;
    /**
     * Component Name
     * @type {string}
     * @memberof RecipeComponentVersion
     */
    componentName: string;
    /**
     * Component Version ID
     * @type {string}
     * @memberof RecipeComponentVersion
     */
    componentVersionId: string;
    /**
     * Component Version Name
     * @type {string}
     * @memberof RecipeComponentVersion
     */
    componentVersionName: string;
    /**
     * Component Version Type (MAIN|HELPER)
     * @type {string}
     * @memberof RecipeComponentVersion
     */
    componentVersionType: string;
    /**
     * Order that the component will run in the recipe
     * @type {number}
     * @memberof RecipeComponentVersion
     */
    order: number;
}

/**
 * Check if a given object implements the RecipeComponentVersion interface.
 */
export function instanceOfRecipeComponentVersion(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "componentId" in value;
    isInstance = isInstance && "componentName" in value;
    isInstance = isInstance && "componentVersionId" in value;
    isInstance = isInstance && "componentVersionName" in value;
    isInstance = isInstance && "componentVersionType" in value;
    isInstance = isInstance && "order" in value;

    return isInstance;
}

export function RecipeComponentVersionFromJSON(json: any): RecipeComponentVersion {
    return RecipeComponentVersionFromJSONTyped(json, false);
}

export function RecipeComponentVersionFromJSONTyped(json: any, ignoreDiscriminator: boolean): RecipeComponentVersion {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'componentId': json['componentId'],
        'componentName': json['componentName'],
        'componentVersionId': json['componentVersionId'],
        'componentVersionName': json['componentVersionName'],
        'componentVersionType': json['componentVersionType'],
        'order': json['order'],
    };
}

export function RecipeComponentVersionToJSON(value?: RecipeComponentVersion | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'componentId': value.componentId,
        'componentName': value.componentName,
        'componentVersionId': value.componentVersionId,
        'componentVersionName': value.componentVersionName,
        'componentVersionType': value.componentVersionType,
        'order': value.order,
    };
}